Everyone's Work
Added everyone's work to the new URP branch bit by bit to avoid any and all conflicts. The only thing left to bring over here are the scripts on enemy AI.

+++ Assets/Scripts/Components/Health.cs
@@ -7,7 +7,7 @@ public class Health : MonoBehaviour
     public float m_MaxHealth;
     private float m_CurrentHealth;
     public bool isEnemy;
-    public float pointsOnDeath;
+    public int pointsOnDeath;
     public bool isDebugging;
 
     // Start is called before the first frame update
